Handles User Object Validation #142
Added in validation for training options and tests as well.

Updates the readme

Updates the readme again and uses a property for `validatesTrainOpts` instead of a getter/setter

Updated the things

I updated the `isStrict` property's name to be longer but more descriptive

I reverted the timeout for async bitwise tests

I moved the `validateTrainingOptions` method to be static

Updated readme and got names a little better.

Updates the things