fix: TrainStream, CrossValidate, and Typescript
Overhauled the internals of TrainStream and CrossValidate so they are decoupled from the networks themselves, as well as being able to restore or store CrossValidate from json.

Added examples for CrossValidate and TrainStream, and as well one for learning math, to test the typescript type file.

Too I added missing typescript configs and examples.

*